requirements.txt
sourcetypes3>=0.1.0
tree_sitter
orjson
pyjson5
psutil